requirements-server.txt
uvicorn==0.38.0
websockets==15.0.1
psycopg2-binary==2.9.11
asyncio==4.0.0